% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porechop.R
\name{read_porechop_log}
\alias{read_porechop_log}
\title{Read useful information from a Porechop log file}
\usage{
read_porechop_log(file)
}
\arguments{
\item{file}{Path to Porechop (ABI) log file}
}
\description{
Read useful information from a Porechop log file
}
